102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,30 +1,24 @@
 #include <stdio.h>
 
-/**
- * main - Entry point
- *
- * Description: Prints the first 50 Fibonacci numbers
- * Return: 0 on success
- */
 int main(void)
 {
-	unsigned long long fib1 = 1; /* First Fibonacci number */
-	unsigned long long fib2 = 2; /* Second Fibonacci number */
-	unsigned long long fib3; /* Variable to store the next Fibonacci number */
-	int count; /* Counter variable */
+    unsigned long fib1 = 1;  /* First Fibonacci number */
+    unsigned long fib2 = 2;  /* Second Fibonacci number */
+    unsigned long fib3;      /* Variable to store the next Fibonacci number */
+    int count;
 
-	printf("%llu, %llu", fib1, fib2); /* Print the first two Fibonacci numbers */
+    printf("%lu, %lu", fib1, fib2);  /* Print the first two Fibonacci numbers */
 
-	for (count = 3; count <= 50; count++)
-	{
-		fib3 = fib1 + fib2;
-		printf(", %llu", fib3);
-		fib1 = fib2;
-		fib2 = fib3;
-	}
+    for (count = 3; count <= 50; count++)
+    {
+        fib3 = fib1 + fib2;
+        printf(", %lu", fib3);
+        fib1 = fib2;
+        fib2 = fib3;
+    }
 
-	printf("\n");
+    printf("\n");
 
-	return 0;
+    return 0;
 }
